+++ src/woo_search/search_index/api/serializers/publications.py
@@ -1,5 +1,6 @@
 from django.utils.translation import gettext_lazy as _
 
+from drf_spectacular.utils import extend_schema_serializer
 from rest_framework import serializers
 
 from ...typing import DocumentIndexType
@@ -20,6 +21,7 @@ class NestedTopicSerializer(serializers.Serializer):
     officiele_titel = serializers.CharField(max_length=255)
 
 
+@extend_schema_serializer(deprecate_fields=("identifier",))
 class DocumentSerializer(serializers.Serializer):
     uuid = serializers.CharField()
     publicatie = serializers.CharField(
@@ -43,7 +45,7 @@ class DocumentSerializer(serializers.Serializer):
         required=False,
         many=True,
         allow_null=True,
-        default=[],
+        default=list,
     )
     publisher = NestedPublisherSerializer(
         help_text=_(
@@ -56,6 +58,20 @@ class DocumentSerializer(serializers.Serializer):
         required=False,
         allow_blank=True,
         default="",
+    )
+    identifiers = serializers.ListField(
+        help_text=_("The document identifiers attached to this document."),
+        child=serializers.CharField(
+            max_length=255,
+            help_text=_(
+                "An identifier specific to this document. Note that multiple "
+                "documents can share identifiers, as additional context is "
+                "required to uniquely identify it, but this context "
+                "is deliberately not indexed."
+            ),
+        ),
+        required=False,
+        default=list,
     )
     officiele_titel = serializers.CharField(max_length=255)
     verkorte_titel = serializers.CharField(
@@ -113,6 +129,11 @@ class DocumentIndexSerializer(DocumentSerializer):
     def validate(self, attrs: DocumentIndexType):
         download_url = attrs["download_url"]
         file_size = attrs["file_size"]
+        identifier = attrs["identifier"]
+        identifiers = attrs["identifiers"]
+
+        if not identifiers and identifier:
+            attrs["identifiers"].append(identifier)
 
         if download_url and file_size is None:
             raise serializers.ValidationError(
@@ -144,7 +165,21 @@ class PublicationSerializer(serializers.Serializer):
         required=False,
         many=True,
         allow_null=True,
-        default=[],
+        default=list,
+    )
+    identifiers = serializers.ListField(
+        help_text=_("The publication identifiers attached to this publication."),
+        child=serializers.CharField(
+            max_length=255,
+            help_text=_(
+                "An identifier specific to this publication. Note that multiple "
+                "publications can share identifiers, as additional context is "
+                "required to uniquely identify it, but this context "
+                "is deliberately not indexed."
+            ),
+        ),
+        required=False,
+        default=list,
     )
     officiele_titel = serializers.CharField(max_length=255)
     verkorte_titel = serializers.CharField(